requirements.txt
alabaster==0.7.16
Babel==2.14.0
cachetools==5.3.3
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
colorama==0.4.6
coverage==7.5.0
distlib==0.3.8
docutils==0.20.1
filelock==3.14.0
idna==3.7
imagesize==1.4.1
iniconfig==2.0.0
Jinja2==3.1.3
MarkupSafe==2.1.5
mypy==1.10.0
mypy-extensions==1.0.0
packaging==24.0
platformdirs==4.2.1
pluggy==1.5.0
protobuf==5.26.1
Pygments==2.17.2
pyproject-api==1.6.1
pytest==8.2.0
pytest-cov==5.0.0
pytest-sugar==1.0.0
requests==2.31.0
ruff==0.4.2
snowballstemmer==2.2.0
Sphinx==7.3.7
sphinx-rtd-theme==2.0.0
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
tox==4.15.0
typing==3.7.4.3
typing_extensions==4.11.0
urllib3==2.2.1
virtualenv==20.26.1
xrpl-py==2.5.0
grpcio==1.64.0;python_version<"3.13"
grpcio==1.66.2;python_version>="3.13"
grpcio-tools==1.64.0;python_version<"3.13"
grpcio-tools==1.66.2;python_version>="3.13"
GitPython==3.1.43
docker==7.1.0
loguru==0.7.2
base58~=2.1.1
cryptography~=42.0.7
PyYAML~=6.0.1
typeguard~=4.3.0

grpcio~=1.66.2
deap~=1.4.3